@@ -4,14 +4,13 @@
   "description": "CRUD interface generator for Replicache",
   "homepage": "https://github.com/rocicorp/rails",
   "repository": "github:rocicorp/rails",
+  "type": "module",
   "main": "out/index.js",
   "scripts": {
     "format": "prettier --write 'src/*.{js,jsx,json,ts,tsx,html,css,md}' '*.{cjs,js,jsx,json,ts,tsx,html,css,md}'",
     "check-format": "prettier --check 'src/*.{js,jsx,json,ts,tsx,html,css,md}' '*.{cjs,js,jsx,json,ts,tsx,html,css,md}'",
     "lint": "eslint --ext .ts,.tsx,.js,.jsx src/",
-    "build-cjs": "tsc --module commonjs --outDir out",
-    "build-esm": "tsc --module ESNext --outDir out && mv out/index.js out/index.mjs",
-    "build": "npm run build-esm && npm run build-cjs",
+    "build": "tsc --module ESNext --outDir out",
     "prepack": "npm run lint && npm run test && npm run build",
     "test": "web-test-runner \"src/**/*.test.{jsx,js,ts,tsx}\""
   },
@@ -30,24 +29,23 @@
   "devDependencies": {
     "@esm-bundle/chai": "^4.3.4-fix.0",
     "@rocicorp/eslint-config": "^0.3.0",
-    "@rocicorp/logger": "^2.2.0",
+    "@rocicorp/logger": "^4.1.0",
     "@rocicorp/prettier-config": "^0.1.1",
     "@snowpack/web-test-runner-plugin": "^0.2.2",
     "@types/chai": "^4.2.22",
     "@types/mocha": "^9.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.3.1",
-    "@typescript-eslint/parser": "^5.18.0",
     "@web/test-runner": "^0.13.18",
     "nanoid": "^3.3.4",
-    "replicache": ">=10.0 <13.0",
+    "replicache": "^12.2.2",
     "snowpack": "^3.8.8",
-    "typescript": "^4.4.3",
+    "typescript": "^4.9.5",
     "zod": "^3.21.4"
   },
   "files": [
+    "out/index.d.ts",
     "out/index.js",
-    "out/index.mjs",
-    "out/index.d.ts"
+    "out/index.d.ts.map",
+    "out/index.js.map"
   ],
   "eslintConfig": {
     "extends": "@rocicorp/eslint-config"
