@@ -2,4 +2,4 @@
 // See all supported options: https://www.snowpack.dev/reference/configuration
 
 /** @type {import("snowpack").SnowpackUserConfig } */
-module.exports = {};
+export default {};
